pyproject.toml
[tool.poetry]
name = "solar-sensor-server"
version = "0.1.0"
description = ""
authors = ["Gabe Hassler"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
paho-mqtt = "^1.6.1"
python-dotenv = "^1.0.0"
pandas = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
